Added Muncie (Starflower) source code
ComicServer brings the best part of the newspaper right to your pocket. With Muncie, the homepage displays up to 8 comics of your choosing. Customization has been a priority, so we’ve added 4 brand-new themes for enhanced user experience. The website has a completely refreshed UI, and we can’t wait for you to try it out.